#1475 - implement patch to allow more general BCs in a cell-based simulation with PDEs; add a flag telling the BC container whether or not to delete its BCs
git-svn-id: https://chaste.cs.ox.ac.uk/svn/chaste/trunk@11816 b5f69653-2bf5-0310-8201-b0dafb8e1405

+++ pde/src/common/BoundaryConditionsContainerImplementation.hpp
@@ -39,8 +39,8 @@ along with Chaste. If not, see <http://www.gnu.org/licenses/>.
 #include "HeartEventHandler.hpp"
 
 template<unsigned ELEMENT_DIM, unsigned SPACE_DIM, unsigned PROBLEM_DIM>
-BoundaryConditionsContainer<ELEMENT_DIM,SPACE_DIM,PROBLEM_DIM>::BoundaryConditionsContainer()
-            : AbstractBoundaryConditionsContainer<ELEMENT_DIM,SPACE_DIM,PROBLEM_DIM>()
+BoundaryConditionsContainer<ELEMENT_DIM,SPACE_DIM,PROBLEM_DIM>::BoundaryConditionsContainer(bool deleteConditions)
+            : AbstractBoundaryConditionsContainer<ELEMENT_DIM,SPACE_DIM,PROBLEM_DIM>(deleteConditions)
 {
     mLoadedFromArchive = false;
 
@@ -74,7 +74,10 @@ BoundaryConditionsContainer<ELEMENT_DIM,SPACE_DIM,PROBLEM_DIM>::~BoundaryConditi
                 //Leave the zero boundary condition until last
                 if (neumann_iterator->second != mpZeroBoundaryCondition)
                 {
-                    delete neumann_iterator->second;
+                    if (this->mDeleteConditions)
+                    {
+                        delete neumann_iterator->second;
+                    }
                 }
             }
             neumann_iterator++;
@@ -84,8 +87,10 @@ BoundaryConditionsContainer<ELEMENT_DIM,SPACE_DIM,PROBLEM_DIM>::~BoundaryConditi
 
     delete mpZeroBoundaryCondition;
 
-    this->DeleteDirichletBoundaryConditions(deleted_conditions);
-
+    if (this->mDeleteConditions)
+    {
+        this->DeleteDirichletBoundaryConditions(deleted_conditions);
+    }
 }
 
 template<unsigned ELEMENT_DIM, unsigned SPACE_DIM, unsigned PROBLEM_DIM>
